IVector2; tons of Vector2X changes to make IVector2 be useful

+++ src/pl/shockah/glib/state/transitions/TransitionThroughColor.java
@@ -52,7 +52,7 @@ public class TransitionThroughColor extends Transition {
 	public void render(Graphics g) {
 		Graphics.setColor(color.setAlpha(alpha));
 		g.toggleAbsolute();
-		g.draw(new Rectangle(new Vector2d(),State.get().getDisplaySize().toDouble()));
+		g.draw(new Rectangle(new Vector2d(),State.get().getDisplaySize()));
 		g.toggleAbsolute();
 	}
